Refactor cli module into a package
