+++ src/greynoise/cli/parser.py
@@ -1,102 +1,21 @@
-"""GreyNoise command line Interface."""
+"""CLI parser."""
 
-import json
 import os
 import sys
 
-from argparse import ArgumentParser, ArgumentTypeError
-from datetime import datetime
-from xml.dom.minidom import parseString
-
-from dicttoxml import dicttoxml
+from argparse import ArgumentParser
 
 from greynoise.api import GreyNoise
-from greynoise.util import CONFIG_FILE, load_config, save_config, validate_ip
-
-
-def main(argv=None):
-    """Entry point for the greynoise CLI.
-
-    :param argv: Command line arguments
-    :type: list
-
-    """
-    if argv is None:
-        argv = sys.argv[1:]
-
-    args = parse_arguments(argv)
-    result = args.func(args)
-
-    if args.format == "json":
-        output = json.dumps(result)
-    elif args.format == "xml":
-        output = parseString(dicttoxml(result)).toprettyxml()
-
-    print(output)
-
-
-def setup(args):
-    """Configure API key."""
-    config = {"api_key": args.api_key}
-    save_config(config)
-    print("Configuration saved to {!r}".format(CONFIG_FILE))
-
-
-def noise(args):
-    """Get all noise IPs generated by internet scanners, search engines, and worms."""
-    return args.api_client.get_noise(date=args.date)
-
-
-def context(args):
-    """Run IP context query."""
-    return args.api_client.get_context(ip_address=args.ip_address)
-
-
-def quick_check(args):
-    """Run IP quick check query."""
-    return args.api_client.get_noise_status(ip_address=args.ip_address)
-
-
-def multi_quick_check(args):
-    """Run IP multi quick check query."""
-    return args.api_client.get_noise_status_bulk(ip_addresses=args.ip_address)
-
-
-def actors(args):
-    """Run actors query."""
-    return args.api_client.get_actors()
-
-
-def ip_address_parameter(ip_address):
-    """IPv4 parameter passed from the command line.
-
-    :param ip_address: IPv4 address value
-    :type ip_address: str
-    :raises argparse.ArgumentTypeError: if ip_address value is invalid
-
-    """
-    try:
-        validate_ip(ip_address)
-    except ValueError as exception:
-        raise ArgumentTypeError(str(exception))
-
-    return ip_address
-
-
-def date_parameter(date):
-    """Date parameter passed from the command line.
-
-    :param date: Date value
-    :type date: str
-    :raises argparse.ArgumentTypeError: if date value is invalid
-
-    """
-    try:
-        return datetime.strptime(date, "%Y-%m-%d").date()
-    except ValueError:
-        raise ArgumentTypeError(
-            "Invalid date: {!r}. Expected format: YYYY-MM-DD".format(date)
-        )
+from greynoise.cli.parameter import date_parameter, ip_address_parameter
+from greynoise.cli.subcommand import (
+    actors,
+    context,
+    multi_quick_check,
+    noise,
+    quick_check,
+    setup,
+)
+from greynoise.util import load_config
 
 
 def parse_arguments(argv):
