+++ src/greynoise/cli/subcommand.py
@@ -1,0 +1,35 @@
+"""CLI subcommands."""
+
+from greynoise.util import CONFIG_FILE, save_config
+
+
+def setup(args):
+    """Configure API key."""
+    config = {"api_key": args.api_key}
+    save_config(config)
+    print("Configuration saved to {!r}".format(CONFIG_FILE))
+
+
+def noise(args):
+    """Get all noise IPs generated by internet scanners, search engines, and worms."""
+    return args.api_client.get_noise(date=args.date)
+
+
+def context(args):
+    """Run IP context query."""
+    return args.api_client.get_context(ip_address=args.ip_address)
+
+
+def quick_check(args):
+    """Run IP quick check query."""
+    return args.api_client.get_noise_status(ip_address=args.ip_address)
+
+
+def multi_quick_check(args):
+    """Run IP multi quick check query."""
+    return args.api_client.get_noise_status_bulk(ip_addresses=args.ip_address)
+
+
+def actors(args):
+    """Run actors query."""
+    return args.api_client.get_actors()
